requirements.txt
fastapi~=0.55.1
uvicorn~=0.11.5
configargparse~=1.2.3
motor~=2.1.0
pyjwt~=1.7.1
bcrypt~=3.1.7
python-miltipart~=0.0.5
